backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
paddlepaddle==2.5.2
paddleocr==2.7.0.3
opencv-python==4.8.1.78
numpy==1.24.3
python-cors==1.7.0 